model changes

+++ src/main/java/com/fernando9825/alcaldiasvrestapi/models/entity/Municipio.java
@@ -11,7 +11,7 @@ import javax.validation.constraints.Size;
 public class Municipio {
 
     @Id
-    private Long municipioId;
+    private Integer municipioId;
 
     @NotBlank
     @Size(min = 5)
